requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.1
httpx>=0.25.1
